pyproject.toml
[project]
name = "caseflood-interview"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "attrs>=25.2.0",
    "black>=25.1.0",
    "dataclasses>=0.8",
    "dearpygui>=2.0.0",
    "fastapi>=0.115.12",
    "ffmpeg-python>=0.2.0",
    "isort>=6.0.1",
    "langchain-community>=0.3.20",
    "langchain>=0.3.21",
    "manim>=0.19.0",
    "networkx>=3.4.2",
    "numpy>=2.2.3",
    "openai>=1.69.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "ruff>=0.9.10",
    "scipy>=1.15.2",
    "sympy>=1.13.3",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.0",
    "langchain-cerebras>=0.5.0",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "werkzeug>=3.1.3",
    "ffmpeg>=1.4",
    "google-cloud-storage>=3.1.0",
    "python-dotenv>=1.1.0",
    "bs4>=0.0.2",
    "langchain-openai>=0.3.11",
    "langchain-anthropic>=0.3.10",
    "langgraph>=0.3.21",
    "anthropic>=0.49.0",
    "qianfan>=0.4.12.3",
    "google-cloud-logging>=3.11.4",
    "flask-cors>=5.0.1",
    "manim-physics>=0.2.4",
    "azure-storage-blob>=12.25.1",
    "azure-monitor-opentelemetry>=1.6.5",
]
